Reorganised abilities folder
Sorted abilities into folders.
