+++ Tag 5.3/Source/Tag/GameplayAbilities/Abilities/Movement/JumpAbility.h
@@ -3,18 +3,18 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "EIGameplayAbility.h"
-#include "CrouchAbility.generated.h"
+#include "Tag/GameplayAbilities/Abilities/EIGameplayAbility.h"
+#include "JumpAbility.generated.h"
 
 /**
  * 
  */
 UCLASS()
-class TAG_API UCrouchAbility : public UEIGameplayAbility
+class TAG_API UJumpAbility : public UEIGameplayAbility
 {
 	GENERATED_BODY()
 public:
-	UCrouchAbility();
+	UJumpAbility();
 
 	virtual void ActivateAbility(
 		const FGameplayAbilitySpecHandle Handle,
@@ -33,6 +33,12 @@ public:
 		const FGameplayAbilitySpecHandle Handle,
 		const FGameplayAbilityActorInfo* ActorInfo,
 		const FGameplayAbilityActivationInfo ActivationInfo) override;
+
+	virtual void EndAbility(const FGameplayAbilitySpecHandle Handle,
+		const FGameplayAbilityActorInfo* ActorInfo,
+		const FGameplayAbilityActivationInfo ActivationInfo,
+		bool bReplicateEndAbility,
+		bool bWasCancelled) override;
 
 	virtual void CancelAbility(
 		const FGameplayAbilitySpecHandle Handle,
