+++ Tag 5.3/Source/Tag/GameplayAbilities/Abilities/Tagging/SelfTagAbility.h
@@ -3,25 +3,19 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "EIGameplayAbility.h"
-#include "Abilities/GameplayAbility.h"
-#include "TagAbility.generated.h"
-
-class UGAT_ServerWaitForClientTargetData;
-class AGATA_SphereTrace;
-class ATagPlayerController;
-class ATagCharacter;
+#include "Tag/GameplayAbilities/Abilities/EIGameplayAbility.h"
+#include "SelfTagAbility.generated.h"
 
 /**
  * 
  */
 UCLASS()
-class TAG_API UTagAbility : public UEIGameplayAbility
+class TAG_API USelfTagAbility : public UEIGameplayAbility
 {
 	GENERATED_BODY()
-
+	
 public:
-	UTagAbility();
+	USelfTagAbility();
 
 protected:
 	virtual void ActivateAbility(const FGameplayAbilitySpecHandle Handle,
@@ -43,12 +37,8 @@ protected:
 	virtual void InputReleased(const FGameplayAbilitySpecHandle Handle,
 		const FGameplayAbilityActorInfo* ActorInfo,
 		const FGameplayAbilityActivationInfo ActivationInfo) override;
-	
-	void AttemptTag(ATagCharacter* TaggingCharacter, ATagCharacter* TagHitCharacter);
-	void RemoveTagEffect(ATagCharacter* TagCharacter);
-	bool Tag(ATagCharacter* CharacterToTag);
 
-	void TryTag();
+	void TagSelf();
 
 	UPROPERTY(EditDefaultsOnly, Category = "Abilities")
 	TSubclassOf<class UGameplayEffect> TagEffectClass;
@@ -57,25 +47,10 @@ protected:
 	UPROPERTY(EditDefaultsOnly, Category = "Abilities")
 	TSubclassOf<class UGameplayEffect> TagDisabledEffectClass;
 
-	UPROPERTY(EditDefaultsOnly)
-	float TagRange = 150.f;
-	UPROPERTY(EditDefaultsOnly)
-	float TagRadius = 75.f;
-
-	UPROPERTY(EditDefaultsOnly)
-	UAnimMontage* TagMontage;
-	
 	FGameplayTag TaggedGameplayCueTag;
-	FGameplayTag AimingTag;
-	FGameplayTag AimingRemovalTag;
 
 private:
-	UPROPERTY()
-	AGATA_SphereTrace* SphereTraceTargetActor;
-	UFUNCTION()
-	void OnTargetDataReady(const FGameplayAbilityTargetDataHandle& TargetData);
-
-public:
-	float GetTagRange() const { return TagRange; }
-	float GetTagRadius() const { return TagRadius; }
+	FGameplayTag TaggedEffectTag;
+	
+	
 };
